requirements.txt
pandas>=1.5.0
openpyxl>=3.0.10
pyyaml>=6.0
python-dotenv>=0.21.0